refactor: flatten dashboard folder
