+++ app/dashboard/[pitchId]/page.tsx
